acts/requirements.txt
Flask
pymysql
datetime
requests
pybase64
flask-mysql
mysql-connector
